+++ libs/ratanpy/src/ratanpy/ratan/ratan_interpolator.py
@@ -1,0 +1,10 @@
+from abc import ABC, abstractmethod
+
+from ratanpy.ratan.ratan_observation import RatanObservation
+
+
+class RatanInterpolator(ABC):
+
+    @abstractmethod
+    def process(self, observation: RatanObservation):
+        pass