+++ libs/ratanpy/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## [Unreleased]
-- None
+### Added
+- **Fast Acquisition 1-3 GHz:** added borrowed noise interpolator, large nan-gaps channel filter
 
 ## [2.0.7] - 2026-07-10
 ### Changed